requirements.txt
absl-py==2.1.0
arrow==1.3.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
config==0.5.1
dateparser==1.2.0
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.6
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
imageio==2.36.1
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
moment==0.12.1
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.9.20
times==0.7
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.0
